Fix/ Edge browser - add Quota param to skip custom max papers validation (#1964)
* feat: update Column component to handle query quota

The code changes in `Column.js` modify the `Column` component to handle the query quota. Previously, the condition for `quotaNotReached` was based on `p.traverseEdgesCount < customLoad`, but now it also checks if the `query.Quota` is present. If `query.Quota` is present, it checks if `p.traverseEdge` is false, otherwise it falls back to the previous condition.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature) and provides a clear and concise description of the changes made.

* change query param name

---------

Co-authored-by: Melisa Bok <melisabok@gmail.com>